Revert "Revert "Drone: readable test report with a permalink test status page…"
This reverts commit 75ec9409e9621ed31bd04ac0eaa7c849aa001896.
