+++ dashboard/test/ui/test_status.haml
@@ -6,7 +6,6 @@
     %link{:rel => :stylesheet, :type => :"text/css", :href => "test_status.css"}
     %script{src: 'https://cdn.jsdelivr.net/lodash/4.13.1/lodash.min.js'}
     %script{src: 'https://cdn.jsdelivr.net/clipboard.js/1.5.13/clipboard.min.js'}
-    %script{src: './js/set-tab-status-icon.js'}
   %body
     .centered-column
       %div#refresh-controls
