+++ dashboard/test/ui/test_status.js
@@ -1,9 +1,12 @@
 // Scripts that operate the test status page.  We expect the page to be
 // prepopulated with tables listing all of the tests we are going to run
 // (see runner.rb and test_status.haml for how this happens) but the actual
-// status of those tests we retrieve asynchronously from a server API
-// (see test_logs_controller.rb) which in turn gets its information from the
-// uploaded S3 logs and their metadata.
+// status of those tests we retrieve asynchronously in one of two ways:
+// 1. from a server API (see test_logs_controller.rb) which in turn gets its
+//    information from the uploaded S3 logs and their metadata. This mode
+//    is faster.
+// 2. directly fetching the metadata from S3 via HEAD requests, this mode
+//    is slower, but allows us to be hosted as static HTML on S3 w/o a backend.
 
 // Globals are provided on test_status.html
 /* global _, Clipboard, setTabStatusIcon */
@@ -309,7 +312,96 @@ function updateProgressNow() {
 }
 var updateProgress = _.debounce(updateProgressNow, 300, { maxWait: 3000 });
 
-function refresh() {
+// maps HTTP headers returned from S3 by a HEAD request to the same names as test_logs_controller.rb returns
+const s3HeaderToMetadataKey = httpHeaderName =>
+  httpHeaderName === 'x-amz-version-id'
+    ? 'version_id'
+    : httpHeaderName.replace(/^x-amz-meta-/, '').replaceAll('-', '_');
+
+// Fetch test results directly from S3, already stored as object metadata on each *_output.html file
+// This function is used when running as static HTML. In this condition, the test_status_UI.html (etc)
+// pages we're running in will be served from the same S3 bucket, and therefore domain, as the *_output.html
+// files. Therefore, there is no CORS issue doing HEAD requests against them.
+async function fetchMetadataDirectlyFromS3For(browser, featureKey) {
+  const featureFilename = featureKey.replace(/^features_/, '').replace(/\.feature$/, '');
+  const s3Key = `${S3_PREFIX}/${browser}_${featureFilename}${S3_KEY_SUFFIX}`;
+  const url = `https://${S3_BUCKET}.s3.amazonaws.com/${s3Key}`;
+
+  const res = await fetch(url, { method: 'HEAD' });
+  if (!res.ok) {
+    console.error(`HTTP ${res.status}: trying to fetch ${url} for browser=${browser} featureKey=${featureKey}`);
+    return undefined;
+  }
+
+  // Transform S3 HTTP header names into the same metadata keys as test_logs_controller.rb returns
+  const metadata = Object.fromEntries(
+    Array.from(res.headers.entries(), ([httpHeader, v]) => [s3HeaderToMetadataKey(httpHeader), v])
+  );
+  metadata.key = s3Key;
+
+  return metadata;
+}
+
+// Iterate through `tests`, and fetch metadata for each test directly from S3.
+// This function works without a backend, its a fallback in particular to serve
+// our HTML off a static bucket.
+async function fetchMetadataDirectlyFromS3() {
+  // Fetch all test results from S3 in parallel, start them going:
+  const entries = Object.entries(tests).flatMap(([browser, features]) =>
+    Object.keys(features).map(featureKey =>
+      fetchMetadataDirectlyFromS3For(browser, featureKey)
+    )
+  );
+
+  // Now wait for all the parallel fetch() calls to complete, this returns the
+  // same format as `test_logs_controller.rb`, and allows to operate "backendless".
+  return (await Promise.all(entries)).filter(Boolean);
+}
+
+// Fetches test status (pass, fail, etc) from S3 metadata by proxying our call through
+// test_logs_controller.rb. We now support `fetchMetadataDirectlyFromS3`, but this is slower
+// because the browser more heavily restricts the number or parallel requests to
+// the same domain. As a result, we prefer this method.
+async function fetchMetadataFromDashboardAPI(since) {
+  const res = await fetch(`${API_BASEPATH}/${S3_PREFIX}/since/${since}`, {
+    mode: "no-cors",
+  });
+  if (res.ok) {
+    return res.json();
+  } else {
+    throw new Error(`HTTP ${res.status}`);
+  }
+}
+
+// Try to fetch S3 metadata for _output.html files (which contains test status like pass/fail)
+//
+// First we try to fetch it via `test_logs_controller.rb`, this will work when dashboard is running
+// in situations like the test server, or your local dev box.
+//
+// If this fails, we fallback to fetching the S3 metadata directly. This is a little slower,
+// but allows us to support a static-html-only mode, that allows uploading this file to S3, giving
+// us a permalink we can pass back in drone.
+async function fetchMetadata(since) {
+  try {
+    // Try to fetch from test_logs_controller.rb first:
+    return await fetchMetadataFromDashboardAPI(since)
+  } catch (error) {
+    // Looks like test_logs_controller.rb wasn't there, we're probably running as static HMTML uploaded to S3
+    console.warn(`Failed to fetch metadata from dashboard API, falling back to trying direct S3 access (original error: ${error})`);
+
+    // Maybe this file is hosted on a static S3 bucket? Or dashboard isn't running?
+    // Lets try accessing S3 directly. The bucket IS public after all.
+
+    // However, first we disable auto-refresh, because unlike S3 we can't do
+    // a one-op list that includes the last_modified metadata, so loading this
+    // every 10s would be doing hundreds of HTTP GET calls over and over.
+    disableAutoRefresh();
+
+    return await fetchMetadataDirectlyFromS3()
+  }
+}
+
+async function refresh() {
   // Fetches all logs for this branch and maps them to the tests in this run.
   // Passes last modification times to the test objects so they can decide
   // whether to update.
@@ -317,37 +409,25 @@ function refresh() {
     return;
   }
   refreshButton.disabled = true;
-  const ensure = () => {
-    refreshButton.disabled = false;
-  };
   let lastRefreshEpochSeconds = Math.floor(lastRefreshTime.getTime() / 1000);
   let newTime = new Date();
-  fetch(`${API_BASEPATH}/${S3_PREFIX}/since/${lastRefreshEpochSeconds}`, {
-    mode: "no-cors",
-  })
-    .then((response) => {
-      if (response.ok) {
-        return response.json();
+  
+  try {
+    const json = await fetchMetadata(lastRefreshEpochSeconds);
+
+    try {
+      for (object of json) {
+        await refreshIndividualTest(object)
       }
-      throw new Error(
-        `While fetching updates, "${response.url}" returned ${response.status}.`
-      );
-    })
-    .then((json) => {
-      return Promise.all(json.map((object) => refreshIndividualTest(object)))
-        .then(() => {
-          lastRefreshTime = newTime;
-          lastRefreshTimeLabel.textContent =
-            "Updated " + lastRefreshTime.toTimeString();
-        })
-        .catch((error) => {
-          lastRefreshTimeLabel.textContent =
-            "Partially updated at " + newTime.toTimeString();
-          console.warn(error);
-        });
-    })
-    .catch((error) => console.error(error))
-    .then(ensure, ensure);
+      lastRefreshTime = newTime;
+      lastRefreshTimeLabel.textContent = "Updated " + lastRefreshTime.toTimeString();
+    } catch (error) {
+      lastRefreshTimeLabel.textContent = "Partially updated at " + newTime.toTimeString();
+      console.warn(error);
+    }
+  } finally {
+    refreshButton.disabled = false;
+  }
 }
 
 /**
@@ -360,6 +440,7 @@ function refreshIndividualTest(object) {
     return test.setLastModified(object, new Date(object.last_modified));
   }
   // If we can't find the test, we don't care about it.
+  console.warn(`refreshIndividualTest(): skipping test for ${object.key} because none could be found, object=`, object);
   return Promise.resolve();
 }
 
@@ -422,3 +503,157 @@ function toggleHelpText() {
   }
 }
 helpLink.onclick = toggleHelpText;
+
+/**
+ * Changes the icon in your browser tab (replacing the favicon)
+ * to a dynamically-generated status icon of the given color
+ * and progress info.
+ * @param {string} [typeOrColor] Any valid HTML color string, or a
+ *   status string ('pass', 'fail', 'in-progress', 'not-started').
+ *   Defaults to gray (not started).
+ */
+function setTabStatusIcon(typeOrColor) {
+  var GRAY = '#aaa';
+  var BLUE = '#3A84CB';
+  var GREEN = '#7FBA00';
+  var RED = '#C40000';
+  var DEFAULT_COLOR = GRAY;
+  var STATUS_COLORS = {
+    'fail': RED,
+    'in-progress': BLUE,
+    'not-started': GRAY,
+    'pass': GREEN
+  };
+  var STATUS_ICONS = {
+    'fail': drawX,
+    'pass': drawCheckmark
+  };
+
+  var LINK_ELEMENT_ID = 'tab-status-icon-link';
+
+  /**
+   * Replace the favicon reference in the document head with a new one.
+   */
+  function setStatusIcon(imageUrl) {
+    var head = document.getElementsByTagName('head')[0];
+
+    var oldLink = document.getElementById(LINK_ELEMENT_ID);
+    if (oldLink) {
+      head.removeChild(oldLink);
+    }
+
+    var newLink = document.createElement('link');
+    newLink.id = LINK_ELEMENT_ID;
+    newLink.rel = 'shortcut icon';
+    newLink.href = imageUrl;
+    head.appendChild(newLink);
+  }
+
+  /**
+   * Generates a PNG dataURL for a status icon displaying the
+   * given color and progress.
+   * @param {string} [typeOrColor] Any valid HTML color string, or a
+   *   status string ('pass', 'fail', 'in-progress', 'not-started').
+   *   Defaults to gray (not started).
+   */
+  function makeStatusIcon(typeOrColor) {
+    if (typeof typeOrColor !== 'string') {
+      typeOrColor = DEFAULT_COLOR;
+    }
+    var color = STATUS_COLORS.hasOwnProperty(typeOrColor) ? STATUS_COLORS[typeOrColor] : typeOrColor;
+
+    var canvas = document.createElement('canvas');
+    canvas.width = canvas.height = 32;
+    var ctx = canvas.getContext('2d');
+
+    drawBackground(ctx, color);
+    if (STATUS_ICONS.hasOwnProperty(typeOrColor)) {
+      STATUS_ICONS[typeOrColor](ctx, color);
+    } else {
+      drawHighlight(ctx);
+    }
+
+    return canvas.toDataURL();
+  }
+
+  /**
+   * Draw a bordered circular background in the given color
+   * @param ctx
+   * @param color
+   */
+  function drawBackground(ctx, color) {
+    // Border
+    ctx.beginPath();
+    ctx.ellipse(16, 16, 16, 16, 0, 0, 2 * Math.PI);
+    ctx.fillStyle = color;
+    ctx.fill();
+
+    ctx.beginPath();
+    ctx.ellipse(16, 16, 14, 14, 0, 0, 2 * Math.PI);
+    ctx.fillStyle = 'white';
+    ctx.fill();
+
+    // Fill
+    ctx.beginPath();
+    ctx.ellipse(16, 16, 13, 13, 0, 0, 2 * Math.PI);
+    ctx.fillStyle = color;
+    ctx.fill();
+  }
+
+  /**
+   * Draw a checkmark (success) icon
+   * @param ctx
+   * @param color
+   */
+  function drawCheckmark(ctx, color) {
+    ctx.beginPath();
+    ctx.moveTo(8, 16);
+    ctx.lineTo(15, 19);
+    ctx.lineTo(29, 6);
+    ctx.lineTo(30, 7);
+    ctx.lineTo(15, 24);
+    ctx.lineTo(7, 17);
+    ctx.closePath();
+    ctx.strokeStyle = color;
+    ctx.lineWidth = 3;
+    ctx.stroke();
+    ctx.fillStyle = 'white';
+    ctx.fill();
+  }
+
+  function drawX(ctx, color) {
+    ctx.beginPath();
+    ctx.moveTo(6, 7);
+    ctx.lineTo(11, 7);
+    ctx.lineTo(17, 15);
+    ctx.lineTo(23, 7);
+    ctx.lineTo(28, 7);
+    ctx.lineTo(18, 16);
+    ctx.lineTo(28, 25);
+    ctx.lineTo(23, 25);
+    ctx.lineTo(17, 17);
+    ctx.lineTo(11, 25);
+    ctx.lineTo(6, 25);
+    ctx.lineTo(16, 16);
+    ctx.closePath();
+    ctx.strokeStyle = color;
+    ctx.lineWidth = 3;
+    ctx.stroke();
+    ctx.fillStyle = 'white';
+    ctx.fill();
+  }
+
+  /**
+   * Draw a "specular highlight."
+   * Used as a default if we don't know what icon to draw.
+   * @param ctx
+   */
+  function drawHighlight(ctx) {
+    ctx.beginPath();
+    ctx.ellipse(22, 10, 4, 4, 0, 0, 2 * Math.PI);
+    ctx.fillStyle = 'white';
+    ctx.fill();
+  }
+
+  setStatusIcon(makeStatusIcon(typeOrColor));
+}
